Crate catbox

Unofficial library implementing catbox.moe’s API in Rust

Separates funtionalities into three modules:

file for uploading and deleting singular files

album for album operations with existing files on Catbox

litter for uploading temporary files to Litterbox

See https://catbox.moe/faq.php for allowed filetypes and content, as well as other questions.

Consider donating via https://www.patreon.com/catbox to help with server costs.

Modules

album

Functions for handling albums through Catbox’s API

file

Functions for handling file upload and deletion through Catbox’s API

litter

Functions for handling temporary file upload through Litterbox’s API